CompadreAnalyticsService/service/requirements.txt
pandas~=2.1.1
numpy~=1.26.0
keras~=2.14.0
tensorflow~=2.14.0
scikit-learn~=1.3.1
xgboost~=2.0.0
django
djangorestframework
boto3
requests
statsmodels